chore: rename fetch API component file
