+++ starters/vue3-apollo-quasar/src/components/FetchExample/FetchExample.vue
