+++ starters/vue3-apollo-quasar/src/components/FetchExample/index.ts
@@ -1,0 +1,1 @@
+export { default as FetchExample } from './FetchExample.vue';
