--- starters/vue3-apollo-quasar/src/components/FetchMessage/index.ts
@@ -1,1 +1,0 @@
-export { default as FetchMessage } from './FetchMessage.vue';
