+++ starters/vue3-apollo-quasar/src/pages/ApiExample.vue
@@ -2,7 +2,7 @@
   <div>
     <h4 class="text-center">Vue3 Apollo Quasar Fetch Data from API</h4>
     <q-separator color="primary" size="4px" />
-    <FetchMessage :message="message" />
+    <FetchExample :message="message" />
   </div>
 </template>
 
@@ -17,7 +17,7 @@ export default defineComponent({
 
 <script lang="ts" setup>
 import { inject, ref } from 'vue';
-import { FetchMessage } from 'src/components';
+import { FetchExample } from 'src/components';
 
 const message = ref('vue3-apollo-quasar starter.dev!');
